Factored out common test infrastructure
